+++ src/main/java/kr/hhplus/be/server/common/application/queue/QueueTokenInfo.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.application.lock.queue;
+package kr.hhplus.be.server.common.application.queue;
 
 public record QueueTokenInfo(String token, Long rank, Long totalWaitMinute, QueueTokenStatus status) {
