+++ src/main/java/kr/hhplus/be/server/common/application/queue/QueueTokenManager.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.application.lock.queue;
+package kr.hhplus.be.server.common.application.queue;
 
 public interface QueueTokenManager {
