+++ src/main/java/kr/hhplus/be/server/common/infrastructure/queue/RedisQueueTokenManager.java
@@ -3,9 +3,9 @@ package kr.hhplus.be.server.common.infrastructure.queue;
 import java.time.Duration;
 import java.util.Optional;
 import java.util.UUID;
-import kr.hhplus.be.server.common.application.lock.queue.QueueTokenInfo;
-import kr.hhplus.be.server.common.application.lock.queue.QueueTokenManager;
-import kr.hhplus.be.server.common.application.lock.queue.QueueTokenStatus;
+import kr.hhplus.be.server.common.application.queue.QueueTokenInfo;
+import kr.hhplus.be.server.common.application.queue.QueueTokenManager;
+import kr.hhplus.be.server.common.application.queue.QueueTokenStatus;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
